refactor(hardware/gyro): separate IMUs and single axis gyros
