+++ src/main/java/net/frc5183/librobot/hardware/gyro/single/SingleAxisGyroscope.java
@@ -1,0 +1,10 @@
+package net.frc5183.librobot.hardware.gyro.single;
+
+import edu.wpi.first.math.geometry.Rotation2d;
+
+public abstract class SingleAxisGyroscope {
+    // todo
+    public enum Axis {
+        YAW, PITCH, ROLL
+    }
+}
